smart-pdf-chat/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==1.10.13
langchain-core==0.1.21
langchain==0.1.0
langchain-community==0.0.20
langchain-google-genai==0.0.11
google-generativeai==0.4.1
pypdf2==3.0.1
python-dotenv==1.0.1
faiss-cpu==1.7.4
numpy==1.26.4
sentence-transformers==2.5.1
torch==2.2.1 